+++ backend/src/main/java/com/example/backend/repository/member/MemberAuthorityRepository.java
@@ -1,4 +1,4 @@
-package com.example.backend.repository;
+package com.example.backend.repository.member;
 
 import com.example.backend.entity.member.MemberAuthority;
 import org.springframework.data.jpa.repository.JpaRepository;
