+++ backend/src/main/java/com/example/backend/repository/member/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.example.backend.repository;
+package com.example.backend.repository.member;
 
 import com.example.backend.entity.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
