+++ backend/src/main/java/com/example/backend/service/MemberCreateService.java
@@ -6,9 +6,9 @@ import com.example.backend.entity.member.Member;
 import com.example.backend.entity.member.MemberAuthority;
 import com.example.backend.oauth2.OAuth2Provider;
 import com.example.backend.oauth2.dto.UserProfileDto;
-import com.example.backend.repository.AuthorityRepository;
-import com.example.backend.repository.MemberAuthorityRepository;
-import com.example.backend.repository.MemberRepository;
+import com.example.backend.repository.member.AuthorityRepository;
+import com.example.backend.repository.member.MemberAuthorityRepository;
+import com.example.backend.repository.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
